speed up of test_big_witness_transaction by reusing of CTransaction.
test_big_witness_transaction was spending most of time converting CMutableTransaction to Transaction, which involves recalculating hashes. This means the time complexity for providing witnesses for the 4500 inputs had a O(n2) component. This change replaces the call to SignSignature (which always creates a new CTransaction) with calls to lower level functions ProduceSignature and UpdateTransaction, which allows for reuse of a single CTransaction. This reduces execution time in about 70%.